combine two files
